+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- K8s.Operation `query_params` map. `labelSelector` 
+- Deprecated HTTPoison options being passed to K8s.Client.Runner.base
+- K8s.Operation.put_query_param/3 to add query parameters by key 
+- K8s.Operation.get_query_param/3 to get a query parameter by key
+
 ### Changed
 - Refactored old references to `cluster_name` to `conn`